pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "dbx-cv-client"
version = "0.1.0"
description = "CLI client for streaming data to Databricks ingestion tables"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "aiohttp",
  "azure-identity",
  "azure-keyvault-secrets",
  "databricks-zerobus-ingest-sdk",
  "grpcio-tools>=1.60.0,<2.0",
  "numpy",
  "opencv-python-headless",
  "typer",
]

[project.scripts]
dbx-cv-client = "dbx_cv_client.cli:app"

[project.optional-dependencies]
dev = ["pytest"]

[tool.uv]
package = true
